requirements.txt
asyncio
aiofiles>=23.1.0
aioshutil
anytree
apscheduler
aioaria2
aioqbt
beautifulsoup4
cinemagoer
cloudscraper
cryptography>=41.0.0
codext>=1.15.6
dnspython
ElectroCrypto
kurigram
fastapi
feedparser
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]
python-multipart
PyJWT
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
gunicorn
httpx
jinja2
langcodes
lxml
natsort
pycountry
pillow
psutil
pymongo
pyshorteners
python-magic
pytz
PyMuPDF
pytdbot[tdjson]
reportlab>=4.0.0
requests
speedtest-cli
telegraph
tenacity
toml
uvicorn
uvloop
vt-py>=0.17.5
xattr
yt-dlp[default,curl-cffi]
git+https://github.com/shabbirmahmud/streamrip.git@dev
git+https://github.com/DraftKinner/zotify.git@dev
